# inline Sigma functions