Before (SHA256)
```
40, 3162, 6513.72, 0.4854368932038835
41, 3868, 6769, 0.5714285714285714
42, 4732, 6719.44, 0.704225352112676
43, 5788, 6771.96, 0.8547008547008547
44, 7079, 6795.84, 1.0416666666666667
45, 8660, 6841.4, 1.2658227848101267
46, 10593, 6885.45, 1.5384615384615385
47, 12957, 6867.21, 1.8867924528301887
48, 15849, 6815.07, 2.3255813953488373
49, 19387, 6785.45, 2.857142857142857
50, 23714, 6742.21568627451, 3.5172413793103448
51, 29007, 6892.752475247525, 4.208333333333333
52, 35481, 6889.514563106796, 5.15
53, 43401, 6875.40594059406, 6.3125
54, 53088, 6833.108910891089, 7.769230769230769
55, 64938, 6803.028571428571, 9.545454545454545
56, 79433, 6808.542857142857, 11.666666666666666
57, 97163, 6801.41, 14.285714285714286
58, 118850, 6856.7307692307695, 17.333333333333332
59, 145378, 6608.090909090909, 22
60, 177828, 6905.941747572816, 25.75
61, 217520, 6851.023622047244, 31.75
62, 266073, 6881.198275862069, 38.666666666666664
63, 325462, 6875.957746478874, 47.333333333333336
64, 398107, 6805.247863247863, 58.5
65, 486968, 6858.704225352113, 71
66, 595662, 6618.466666666666, 90
67, 728618, 6873.754716981132, 106
68, 891251, 6855.776923076923, 130
69, 1090184, 6899.898734177215, 158
70, 1333521, 6873.819587628866, 194
71, 1631173, 6740.384297520661, 242
72, 1995262, 6786.60544217687, 294
73, 2440619, 6874.98309859155, 355
74, 2985383, 6831.54004576659, 437
75, 3651741, 6890.077358490566, 530
76, 4466836, 6850.9754601227, 652
77, 5463865, 6872.786163522012, 795
78, 6683439, 6883.0473738414, 971
79, 8175230, 6801.356073211315, 1202
80, 10000000, 6872.852233676976, 1455
```

After
```
40, 3162, 17232.9, 0.1834862385321101
41, 3868, 17328.64, 0.22321428571428573
42, 4732, 17319.12, 0.273224043715847
43, 5788, 17421.88, 0.33222591362126247
44, 7079, 17555.92, 0.4032258064516129
45, 8660, 17493.2, 0.49504950495049505
46, 10593, 17584.38, 0.6024096385542169
47, 12957, 17103.24, 0.7575757575757576
48, 15849, 17592.39, 0.9009009009009009
49, 19387, 17642.17, 1.098901098901099
50, 23714, 17785.5, 1.3333333333333333
51, 29007, 17694.27, 1.639344262295082
52, 35481, 17740.5, 2
53, 43401, 17618.227722772277, 2.4634146341463414
54, 53088, 17696, 3
55, 64938, 17533.26, 3.7037037037037037
56, 79433, 17132.607843137255, 4.636363636363637
57, 97163, 17750.93269230769, 5.473684210526316
58, 118850, 17650.9900990099, 6.733333333333333
59, 145378, 17499.203703703704, 8.307692307692308
60, 177828, 17782.8, 10
61, 217520, 17797.090909090908, 12.222222222222221
62, 266073, 17738.2, 15
63, 325462, 17752.47272727273, 18.333333333333332
64, 398107, 17772.633928571428, 22.4
65, 486968, 17086.59649122807, 28.5
66, 595662, 17692.930693069306, 33.666666666666664
67, 728618, 17771.170731707316, 41
68, 891251, 17648.534653465347, 50.5
69, 1090184, 17726.56910569106, 61.5
70, 1333521, 17780.28, 75
71, 1631173, 17730.141304347828, 92
72, 1995262, 17053.521367521367, 117
73, 2440619, 17814.73722627737, 137
74, 2985383, 17664.988165680472, 169
75, 3651741, 17726.898058252427, 206
76, 4466836, 17796.159362549803, 251
77, 5463865, 17568.69774919614, 311
78, 6683439, 17822.504, 375
79, 8175230, 17810.958605664488, 459
80, 10000000, 17730.49645390071, 564
```

Looking at 250% improvement in performance.
Probably won't be consistent,  but,  I doubt there will ever be a performance *loss* given the simplicity of the change.