@dcousens 2015-08-27
Will do this for SHA256, hold up