@dcousens 2015-08-27
Performance improvement was basically 0% for SHA512,  but,  for consistency I kept the changes.