@dominictarr 2015-08-28
Funny that V8 couldn't figure this out on it's own! these are named functions and have no free variables. it would be straightforward to have a preprocessor that made substitutions like this.