@dominictarr 2015-08-28
I'm seeing a much smaller difference. from about 12k bytes/second to 14k. this is with iojs@2.3.1 what version are you using?