@dcousens 2015-08-28
@dominictarr for these tests? `node v0.12.7`