@dominictarr 2015-08-28
okay so my result for your branch is about the same, but my master branch is 2x as fast as you. what commit are you testing against?