@dominictarr 2015-08-28
installing nvm so i can test with other versions...