@dcousens 2015-08-28
@dominictarr I was testing before and after.  Literally.  Stash, no stash.