@dominictarr 2015-09-14
This is an improvement I am seeing 11% improvement, but I am just curious about why you are seeing much more divergent results than I am. @dcousens could you run the benchmark again and see if you get similar results.

the evidence I have gathered is still suggesting this is worthwhile merging, but there seems to be something we don't understand going on and that is really bugging me.