@dominictarr 2015-09-14
merged into 2.4.3